pyproject.toml
[project]
name = "her"
version = "0.0.1"
description = "Family-aware voice + text AI assistant — see CLAUDE.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.3.0",
    "sqlite-vec>=0.1.0",
    "structlog>=24.0",
    "tenacity>=9.0",
    # Phase 1 — Voice Channel
    "sounddevice>=0.4",
    "numpy>=1.26",
    "silero-vad>=5.0",
    "faster-whisper>=1.0",
    # Phase 3 — Wake Word: reuses faster-whisper + silero-vad (no extra dep)
    # Phase 3.5 — Presence Channel (orb WebSocket server)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=13",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "black>=24.0",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
